# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AssignmentStatus": {
        "enum": [
          "assigned",
          "in_progress",
          "submitted",
          "grading",
          "results_ready",
          "correcting",
          "completed",
          "withdrawn",
          "stopped"
        ],
        "title": "AssignmentStatus",
        "type": "string"
      },
...
      "Child": {
        "properties": {
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "grade_stage": {
            "title": "Grade Stage",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "nickname": {
            "title": "Nickname",
            "type": "string"
          },
          "ui_language": {
            "default": "en",
            "title": "Ui Language",
            "type": "string"
          }
        },
        "required": [
          "family_id",
          "nickname",
          "grade_stage"
        ],
        "title": "Child",
        "type": "object"
      },
...
      "CreateChildRequest": {
        "properties": {
          "grade_stage": {
            "maxLength": 60,
            "minLength": 1,
            "title": "Grade Stage",
            "type": "string"
          },
          "nickname": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Nickname",
            "type": "string"
          },
          "pin": {
            "pattern": "^\\d{6}$",
            "title": "Pin",
            "type": "string"
          },
          "ui_language": {
            "default": "en",
            "enum": [
              "zh",
              "ja",
              "en"
            ],
            "title": "Ui Language",
            "type": "string"
          }
        },
        "required": [
          "nickname",
          "grade_stage",
          "pin"
        ],
        "title": "CreateChildRequest",
        "type": "object"
      },
...
      "CreateFamilyInvitationRequest": {
        "properties": {
          "email": {
            "maxLength": 254,
            "minLength": 3,
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "CreateFamilyInvitationRequest",
        "type": "object"
      },
      "CreateFamilyRequest": {
        "properties": {
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateFamilyRequest",
        "type": "object"
      },
...
      "Family": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "Family",
        "type": "object"
      },
      "FamilyInvitation": {
        "properties": {
          "accepted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "external_notification_sent": {
            "const": false,
            "default": false,
            "title": "External Notification Sent",
            "type": "boolean"
          },
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "invited_by": {
            "title": "Invited By",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          }
        },
        "required": [
          "family_id",
          "email",
          "invited_by"
        ],
        "title": "FamilyInvitation",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryItem": {
        "properties": {
          "assignment_id": {
            "format": "uuid",
            "title": "Assignment Id",
            "type": "string"
          },
          "attempt_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Id"
          },
          "available_points": {
            "title": "Available Points",
            "type": "number"
          },
          "awarded_points": {
            "title": "Awarded Points",
            "type": "number"
          },
          "child_id": {
            "format": "uuid",
            "title": "Child Id",
            "type": "string"
          },
          "child_nickname": {
            "title": "Child Nickname",
            "type": "string"
          },
          "correction_count": {
            "title": "Correction Count",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/AssignmentStatus"
          },
          "submitted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "assignment_id",
          "attempt_id",
          "child_id",
          "child_nickname",
          "title",
          "status",
          "submitted_at",
          "awarded_points",
          "available_points",
          "correction_count"
        ],
        "title": "HistoryItem",
        "type": "object"
      },
...
      "ManagementPinRequest": {
        "properties": {
          "pin": {
            "pattern": "^\\d{6}$",
            "title": "Pin",
            "type": "string"
          }
        },
        "required": [
          "pin"
        ],
        "title": "ManagementPinRequest",
        "type": "object"
      },
      "ManagementPinStatus": {
        "properties": {
          "configured": {
            "title": "Configured",
            "type": "boolean"
          }
        },
        "required": [
          "configured"
        ],
        "title": "ManagementPinStatus",
        "type": "object"
      },
      "ManagementUnlockResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "ManagementUnlockResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Learning Assessment API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/families": {
      "get": {
        "operationId": "list_families_v1_families_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Family"
                  },
                  "title": "Response List Families V1 Families Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Families",
        "tags": [
          "families"
        ]
      },
      "post": {
        "operationId": "create_family_v1_families_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFamilyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Family"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Family",
        "tags": [
          "families"
        ]
      }
    },
    "/v1/families/{family_id}/children": {
      "get": {
        "operationId": "list_children_v1_families__family_id__children_get",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Child"
                  },
                  "title": "Response List Children V1 Families  Family Id  Children Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Children",
        "tags": [
          "families"
        ]
      },
      "post": {
        "operationId": "create_child_v1_families__family_id__children_post",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Child"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Child",
        "tags": [
          "families"
        ]
      }
    },
    "/v1/families/{family_id}/invitations": {
      "post": {
        "operationId": "create_family_invitation_v1_families__family_id__invitations_post",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFamilyInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyInvitation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Family Invitation",
        "tags": [
          "families"
        ]
      }
    },
    "/v1/families/{family_id}/management-pin": {
      "get": {
        "operationId": "get_management_pin_status_v1_families__family_id__management_pin_get",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagementPinStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Management Pin Status",
        "tags": [
          "families"
        ]
      },
      "put": {
        "operationId": "set_management_pin_v1_families__family_id__management_pin_put",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Management-Unlock",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Management-Unlock"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagementPinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Management Pin",
        "tags": [
          "families"
        ]
      }
    },
    "/v1/families/{family_id}/management-unlock": {
      "post": {
        "operationId": "create_management_unlock_v1_families__family_id__management_unlock_post",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagementPinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagementUnlockResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Management Unlock",
        "tags": [
          "families"
        ]
      }
    },
...
    "/v1/history/families/{family_id}": {
      "get": {
        "operationId": "list_family_history_v1_history_families__family_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HistoryItem"
                  },
                  "title": "Response List Family History V1 History Families  Family Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Family History",
        "tags": [
          "history"
        ]
      }
    },
...
  }
}